review: rename crypto to encryption
