+++ src/main/java/io/numaproj/kafka/encryption/EnvelopeDecryptionFactory.java
@@ -1,4 +1,4 @@
-package io.numaproj.kafka.crypto;
+package io.numaproj.kafka.encryption;
 
 import java.time.Clock;
 import java.util.Properties;
